+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 * added `Player.allowScroll`
 
+* added `Utils.time.timeRegexp`
+
 ## 1.1.1 (October 20, 2019)
 
 * fix typings for `utils/animation/replay`